Partial controller class added, schedule bugfix
OnCallRotation controller is now a partial class, our added code is
under group code folder. Regenerating the controller will no longer wipe
out our code. Fixed a bug in the schedule where the rotation length of
the application was being treated as a number of days rather than the
intended number of weeks of a rotation.