Calendar now renders events
Created additional HomeController class as a partial to provide a list
of JSON objects to calendar.

Moved calendarEvents.js script directly into Index.cshtml which seems to
have fixed the lack of rendering. I don't understand why this fixed it.
The view is pretty ugly, we'll need to work on cleaning it up a bit.