Improved Schedule
Schedule algorithm now accounts for the number of primary rotations an
employee has had. When the end of the employee list is reached, it is
reordered to take into account new primary rotation counts. This means
that employees will be appropriately alternated between primary and
secondary on-call.

This required several new functions and a struct in the Behavior file.
The generate schedule function has gotten pretty big again, despite
Raphael and I refactoring it a bit earlier.