Seeded holidays for 2015, added holiday retrieval
When generating a schedule, the scheduler constructor checks the
database for the last holiday in the table. If it is within
approximately 6 months the constructor starts a new thread to get the
upcoming year's holidays. The thread is joined at the end of the
generate schedule function to avoid race conditions to the table in
other processes.