Creating holiday rotations updates HasPaidHoliday
Creating holiday rotations now updates the HasPaidHoliday lookup table
to accurately reflect the holidays associated with that rotation.

Tweaked the CreateRotationWithHoliday function to return to old state in
order to better balance rotation assignments after a holiday. It
performs a check before returning to make sure the employee assigned to
the holiday isn't about to be assigned to the next rotation.