Removed extraneous null checks
I was making a bad assumption when making new rotations by thinking that
a rotation might already have holidays tied to it. I moved the
initialization of the empty holidays list to the createRotation method
and dropped the null checks on all the for loops that add holidays to
the list. The related portions of code should be a little easier to read
now.