Fixes to RegenerateSchedule
Regenerate schedule in the controller now returns an existing view
rather than crashing - I missed this in testing because I copied the
code for regenerate into the existing generate schedule to test. It also
now saves the rotations, an error I missed for the same reason.

Regenerate schedule on the back end now deletes rotations solely based
on starting date to handle rotations that are only partially covered in
a clean way. Any rotation that STARTS in the time span given is deleted
to avoid dealing with half rotations.