Update to regenerate UI and Index
OnCallRotations/Index.cshtml now contains all the code that used to be
in generateschedule.cshtml which means that the download and regenerate
functionality can be accessed simply by navigating to the On Call
Rotations page rather than by being forced to click on Generate Schedule
on the home page, waiting, and altering rotation data in the process.

Also, the Java Script code now passes the begin and end dates to the URL
so that the regenerate function can access them.