Improvements to Scheduling
The scheduling algorithm now also attempts to balance the secondary
rotation count of employees as well as the primary count at the expense
of some runtime speed.

The scheduling algorithm now calculates the mean and standard deviation
of primary rotation counts. If an employee is more than 1 standard
deviation from the mean they will be passed up for the next employee.

This commit is in response to issue #12