Fix to Regenerate Schedule
Setting the start date in regenerate schedule to be in the middle of a 2
week rotation would start a new rotation for that application on the
upcoming Wednesday. If set to a Wednesday in the middle of a rotation,
that meant a new rotation began that day and doubled up rotations for
that application creating a messy offset and a gap in the calendar.

GenerateSchedule now checks for an overlapping rotation before beginning
an application's rotations. If an overlap exists, LastFinalDateByApp is
adjusted to the end of the overlapping rotation thus preventing the
offset and doubling up of rotations.