Better calendar offset, POST for event change
Now posting results to URL rather than refreshing entire page. This
operation is much faster.

Calendar offset is now handled in a more semantically clear way for the
user. The datepicker for start date is incremented by one day to
accurately reflect the date in the database. I am unclear as to why
FullCalendar subtracts a day from both the start and end. This change
negates the need for adjusting offsets in UpdateRotations so this will
be easier to maintain.

HSF application is only supposed to have a 1 week rotation. I changed
that to be more accurate.

Further cleanup for Issue #21